CalculadorDistancias funcionando

+++ src/main/java/edu/fiuba/algo3/modelo/Rango.java
@@ -1,5 +1,5 @@
 package edu.fiuba.algo3.modelo;
 
 public interface Rango {
-    int calcularDemoraPorDistancia(int distancia);
+    int calcularDemoraPorDistancia(double distancia);
 }
